перенес все функции в func создал test, есть вопросы
